Updated openstack/openstack Project: openstack/tempest  2eaa27f7a83d9209b7d78ce6497146aeb0c3ae1a
javelin 2

Create a new tool which can create and verify a set of resources to
be used in place of the existing javelin tooling in grename.

This currently can create:
  tenants, users, objects, images, servers

And can verfiy:
  users, objects, servers

A sample resources.yaml file is included.

Follow on patches will handle the additional create / verify, as well
as destroy operations, and unit tests to meta test the script beyond
the normal javelin runs. However at this point it is useful enough to
consider adding to Tempest to being putting this into grenade.

Part of blueprint:javelin2

Change-Id: I33f1901802c92a2eda2eeb2c225bc7f60d1a186c