Updated openstack/openstack Project: openstack/tempest  568d4ea744107f167bcd9f464c677b7ee25ebbcf
Cleanup docstring in test_l3_agent_scheduler

The docstring was incomplete in saying it required the
l3_agent_scheduler extension to be enabled.  I think it probably meant
to say that l3_agent_scheduler was listed in
network_feature_enabled.api_extensions, but that defaults to 'all' so
rather than explain how that config option works in the docstring, just
say that the l3_agent_scheduler extension is required and assume the
user understands how to configure Tempest.

Change-Id: If504bd6374b83dea24fa92f9ef6c852d76308ad1