Updated openstack/openstack Project: openstack/tempest  57f8b18a2e0bb9e98384606919c17cdfe9eca590
test_aggregates_basic_ops picks a non compute node

test_aggregates host ties to add non-hypervisor node,
(does not runs a nova compute service) to an aggregate.

It can cause failures, if the openstack deployment has a node,
with one of the following services n-net, n-cond, n-sch, n-net,
 but without n-cpu.

It is a similar issue, what we had with the aggregate api test.

Change-Id: Idbe037da73169e0ebce8a8bb5d7652dcc39eb92b
Closing-Bug: #1318578
Related-Bug: #1294511