Updated openstack/openstack Project: openstack/neutron  911e04aaf52634e3c2d7ea6bde25130fed617f09
NVP plugin:fix delete sec group when backend is out of sync

If a security group does not exist on the NVP backend, an error
should not be raised on deletion of the security group.

This patch changes the plugin behavior by deleting the record
from the database and just logging that the security group
was not found on the NVP backend.

Closes-Bug: #1251422

Conflicts:
	neutron/tests/unit/nicira/test_nicira_plugin.py

Change-Id: Ib8adf7a830ff336655fd83ad4118cde641adf284
(cherry picked from commit ed83265bb794d38f6f66d9e5182a9abc6b82ed53)