Updated openstack/openstack Project: openstack/neutron  b2f828048532185caa883cb467a41d6329d79e1a
Kill 'Skipping unknown group key: firewall_driver' log trace

This is done by trying to import the option first. If this
does not work, emit a warning instead as in most cases this is
harmless for a number of reasons: a) the service might not
even need the opt; b) if things do break down the line, we'll
see bigger traces; c) it's not gonna be long for this legacy
quantum/neutron stuff to be removed altogether.

Closes-bug: 1210121

Change-Id: I34917da9cb6117ee1d42140621c742f503279b6b
(cherry picked from commit b5ee49623982530bfb3c3fe2eefb9d8ddb6353bc)