Updated openstack/openstack Project: openstack-dev/devstack  47ae725f1337ba76189604b685ccaec6c7b7bff9
be a little more clever about interface default

the interface default of eth0 has become more wrong since we've
got a generation of Linux distros that don't use eth* as their
interface naming. Instead of that do a slightly more clever, but
probably still wrong thing, of picking an interface that exists
and is a physical interface.

This is still overridable, however won't cause really cryptic
failures in nova-compute because it's trying to use an interface
that doesn't exist *really* late in the network creation.

Change-Id: I93016ff5a25678de76254418155c839a269486a0