create a script to regen gitmodules
the openstack/openstack repo is somewhat out of date in the content
that it contains, especially considering all the new oslo modules.

This change creates a script which goes through config repo and
pulls the integrated-gate* projects out. There are a couple of
manual adds after that point, then create the new .gitmodules.

An updated .gitmodules is included.

Change-Id: I5d833a9fd30ca46356cb59b23aac98d0a3e17ed4