Updated openstack/openstack Project: openstack/tempest  a7d32921fbcdb6877388e00b2676c7c77c767f63
Remove the list_servers_since_test

This test is mostly just a very good way to add new failures. There
are subtleties on the timestamping of servers which means that this
can often return the wrong number of servers. Because of this we
should just delete the test.

If we want something like this we should create a much more sane setup
that includes getting all the servers, determinging some subset that
the filter should return, fetching those, and getting the count.

Change-Id: I5c08ef031043229b76a6800eeaab6d418fc4614f