Updated openstack/openstack Project: openstack/oslo-incubator  e3b52f20367e18c16be950bcc8ebc5f2336ba4c5
Add source code generation to conf.py

The sourcecode index was not generated previously for oslo-incubator.
This patch adds the code snippet, which generates autoindex for the
docstrings in each of the source files i nthe openstack folder. The
code snippet is based on the one that is used for Ceilometer.

The index.rst is modified to contain the correct autoindex reference.

The tox.ini was also updated with a 'docs' section.

Change-Id: I0afe2712246e4735e9ad97dd4bc2f98cd7d85613