Updated openstack/openstack Project: openstack/oslo.i18n  67ac9babdfd3f9350f6dd2ab043519000a853ae0
Check the lazy flag at runtime

Update lazy flag handling to check it at runtime when each message is
translated, instead of once at startup when the translation function is
created. This allows a library to support lazy translation, without
requiring an application that uses the library to support it.

Change-Id: Iae22668119c8d0f5fb9c486436bfc35cdb88ac58