Updated openstack/openstack Project: openstack/tempest  a255783fdcb540a37b9881d7f50de247b3e9b587
Wait for server activations for the updated servers

In test_list_servers_by_changes_since, Tempest tries to get a server
list with changes-since parameter. The test expects we can get all
servers with the parameter, but sometimes we could not get it and the
gate tests failed.
On "create a server" API Nova creates a DB record of the server as
the first step, then updates the record by each step(scheduling,
networking, etc.) during the server creation. If the test runs just
after the DB record creation, the record is not updated and "get a
server list" API with changes-since parameter does not contain the
server. As the result, the test fails.

This patch makes the test wait for server activations to avoid the
non-updated servers.

Change-Id: I380ae2030da794c4f8fcfd6347f9cbe7f75b5c28
Closes-Bug: #1254238