Updated openstack/openstack Project: openstack/trove  4e5693d9f760a2c114a4feb5286ff071a6d2c428
Adds the foundation for datastore capabilities

This allows for capabilities to be set in the mysql database capabilities
table and associated to datastores with the datastore_capabilities
table. This allows users to have datastore specific functionality that
is configurable and able to be served to end users. This commit is for
the groundwork for the capabilities API.

Change-Id: I7153d435bf8c00dc2a874a66c7a2b64e8fbafa09
Partially-Implements: blueprint capabilities