Updated openstack/openstack Project: openstack/neutron  d1a9a9dff93e7edb67d1999aa42433a9c06773da
LBaaS add missing rootwrap filter for route

If one runs the lbaas agent from packages and does not have the l3-agent
installed on the same box as the lbaas agent it will fail to add the
default gw route. This is because it's missing the rootwrap filter for
route which is only present in l3.filters.

Conflicts:
	etc/neutron/rootwrap.d/lbaas-haproxy.filters

Change-Id: I59cd0a549a1f1d1564d139d42daf55d02898bf76
Closes-bug: 1325406
(cherry picked from commit c9a0eaacaad3d6d00867a75b5c2302b3836ee46d)