Updated openstack/openstack Project: openstack/heat  bfa6a2df5e885d6ee1cbd3355031dc539998a780
Fix exceptions in update_stack

Updating an in-progress stack is not supported by Heat.  Previously, a
NotSupported exception would be raised.  This was not consistent with
other actions, such as abandon stack.  Instead, let StackLock.acquire
raise an ActionInProgress exception.  Suspended stacks will still result
in a NotSupported exception being raised if an attempt is made to update
them.

Closes-Bug: #1308109
Change-Id: Ic844b06cd654d7bd5d422bba81d1d8246a5d8dd0