Updated openstack/openstack Project: openstack-dev/devstack  f8d4efc0587ce044cbaa3f27bb1dfbfa1f3c271d
Update setuptools to latest for .dist-info support

Support for .dist-info directories was added in setuptools 0.6.28.

At this moment, Ubuntu Precise 12.04 provides setuptools 0.6.24
which is too old for our needs.

Six is installed from wheel which uses the .dist-info directory.
For six to be found, we need to install setuptools >= 0.6.28.

Updating setuptools to the latest version using pip will provide use
the needed version to make six discoverable.

Closes-bug: #1326811
Change-Id: I761d0aeb2b8b593cee38d512afc8fed6a2d1fe37