Updated openstack/openstack Project: openstack/ceilometer  3d2723ed7d09724e743869d9ba7b9b37692ed340
Refactor tests to remove direct access to test DBManagers

Previously in several tests DBManager was specified explicitly and some
tests tried to use it even if no backend was available.
Now we should not access them directly. Instead we can mark certain
tests as applicable only to certain backend driver.
One will only have to provide an URL for tests to use.

Later this code will be used to integrate jenkins jobs on real backends.
Database URL will be provided by jenkins and tests will figure out which
manager should be used and what tests can run with it.

Related to blueprint sql-unit-tests-on-real-backend

Change-Id: I9910c25fffb971addec9cec33b624e3b8387f310