Updated openstack/openstack Project: openstack/tempest  875b36b1cb913bb1b898a3f7d429f733ecb4c244
Add bug skip for test_container_synchronization()

This commit adds a skip to test_container_synchronization(). This test
currently only gets run during the periodic -all job because of it's
slow tag. It is currently failing 100% of the time and has been in this
state for quite some time. So let's skip it until the underlying bug is
fixed.

Related-Bug: #1317133

Change-Id: Iff71b74a377cf80be21516c546031e784a61f836