Updated openstack/openstack Project: openstack/tempest  4d49e1af2d2e3190be8fd151bdb625923fb76ce1
Place baremetal common methods in base class

This is part 1 of a change to add advanced operations tests for
the ironic baremetal tests. This moves some methods, currently
defined in the BaremetalBasicOpsPXESSH class, up to the base class
where they can be used by both basic and advanced tests. Advanced
test will be added in a separate review.

The ironic state definitions are also moved up to the manager.py
file where they can be easily accessed.

This also fixes a typo and renames BaremetalBasicOptsPXESSH to
BaremetalBasicOpsPXESSH.

Change-Id: I6a63835170548759c3f813339186deb15e0dacc4