Updated openstack/openstack Project: openstack/horizon  07e2fe10faa9814945fc9999ec2dd3468e6bdd6c
APIDictWrapper getitem lookups fail for non str

The APIDictWrapper class within openstack-dashboards/api/base.py can't
handle the "in" operator correctly.

Also, it will fail if a non-string is passed into it's __getitem__
method. (Which is what is called from a dict lookup.)

The get method also fails on non-strings.

Change-Id: Iaad66d3c6deb81d0570b1a8c6ddf580024fa46e6
Closes-Bug: 1326512