Updated openstack/openstack Project: openstack/oslo.db  f9167b0e1e7a5dd0bf63d22f17fca99a995ee560
Make get_session() pass kwargs to a sessionmaker

get_session() method is used to construct a new SQLAlchemy session
instance. Currently, it takes two keyword arguments which are used
to configure the Session instance to be created. Actually, there are
more arguments which can be passed to a sessiomaker instance. oslo.db
should not stay in the way of a developer who wants to use specific
features of SQLAlchemy session: it should handle oslo.db specific
arguments and pass keyword arguments as is when calling the
sessionmaker.

Change-Id: I1f577a30586eafe1295bcbd15a6a881e9f7f65ab