Updated openstack/openstack Project: openstack-dev/devstack  3ce3ee6db9e025df6b76b228a8041caf9c42c8ba
Update again and retry a failed package install

On ubuntu, if we run into an error installing the package, it can be
because of transient upstream errors with repo sync state. Although
retrying the update won't fix it all of the time, it's low cost enough
to be worth a try before we totally give up.

Related-bug: 1286818
Change-Id: I522ac0d0bd8f82dc98f386c89f66c2b743efa525