Updated openstack/openstack Project: openstack/taskflow  0adf714f3a3f01cfc74780b1a35367fd487eb474
Create a considerations section

Instead of having the considerations and best practices that
should be used when developing with this library scattered
across a few different links and pages combine these together
into a single section in the developer documentation. As a
side-effect of this change remove the documenation about the
utils to use and not use and just refer those to use the modules,
classes and APIs described in these documents.

Change-Id: I20e1405cb6ecb654baa29812b98ada7bdc393f6c