Updated openstack/openstack Project: openstack/ceilometer  0600d3877f4184d42c4429f67856d55c07975d0a
Bump hacking to 0.9.x

Hacking 0.9.0 had a minor bug so require 0.9.2 or higher.

In order to keep this patch to just a requirements bump, ignore new and
stricter hacking rules that are being triggered. Fixing up the code and
turning these on is out of scope of this patch and is for future patches.

Change-Id: I64f46c92e49625a519c902d402be94c914c50816