Updated openstack/openstack Project: openstack-dev/devstack  58608a1d27841ceab799a51a7118102e73d3c90d
convert devstack from awk to outfilter

The complexity of the awk script was pretty high, and it would
be good to do it in summaries as well, which starts to get a bit
squirelly. Instead bring over the outfilter.py from grenade and
use it for the timestamping.

Any additional overhead from python should be offset from not
shelling out to date on every line of output.

(cherry picked from commit 62cb563ddc5f6923a9734a96d394ff2c27c320ca)

Conflicts:
	stack.sh

Change-Id: Ic2b86ddba3e7f6520a0fd35599b01143936c6deb